Distances
As it might not be feasible for us to try and get the exact distances of the trips trough any map APIs, we could take a really straight forward way and use the shortest distances from point a to point b for our approximation. This combined with a reasonably simple segmentation of the new york city map, and with additional variables could provide us with a sufficient set of predictors to use for estimation of the duration of the trip. What do you think?